Now using uint8 for optimization

+++ functions.py
@@ -122,7 +122,7 @@ def createMessageACK(syncPattern, SVID, msgID):
 
     for i in range(len(msg)):
         arr_msg[i] = int(msg.pop())
-    arr_msg=arr_msg.astype(int)
+    arr_msg=arr_msg.astype(np.uint8)
 
     return arr_msg
 
@@ -152,7 +152,7 @@ def createMessageNACK(syncPattern, SVID, msgID):
 
     for i in range(len(msg)):
         arr_msg[i] = msg.pop()
-    arr_msg=arr_msg.astype(int)
+    arr_msg=arr_msg.astype(np.uint8)
 
     return arr_msg
